+++ Entities/Order.cs
@@ -8,7 +8,7 @@ namespace Orders.Entities
 
         public required List<Product> Products { get; set; }
         
-        public required Double TotalPrice { get; init; }
+        public required int TotalPrice { get; init; }
 
         public DateTimeOffset CreatedDate { get; init; }
     }
